pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "matscope"
version = "0.2.0"
description = "Diagnostic toolkit for scientific foundation models. Understand what your model learned, where it will fail, and how representations evolve across layers."
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.9"
authors = [
    {name = "Joshua Steier"},
]
keywords = [
    "machine-learning",
    "foundation-models",
    "interpretability",
    "probing",
    "atomistic-simulation",
    "materials-science",
    "representation-learning",
    "composition-projection-decomposition",
    "equivariant-neural-networks",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Chemistry",
    "Topic :: Scientific/Engineering :: Physics",
]

dependencies = [
    "numpy>=1.24",
    "scikit-learn>=1.3",
]

[project.optional-dependencies]
torch = ["torch>=2.0"]
mace = ["mace-torch>=0.3"]
chgnet = ["chgnet>=0.3"]
pyg = ["torch-geometric>=2.4"]
ani = ["torchani>=2.2"]
viz = ["matplotlib>=3.7", "seaborn>=0.12"]
full = [
    "torch>=2.0",
    "mace-torch>=0.3",
    "chgnet>=0.3",
    "torch-geometric>=2.4",
    "torchani>=2.2",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "pandas>=2.0",
    "scipy>=1.11",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1",
    "mypy>=1.5",
    "pre-commit>=3.0",
]

[project.urls]
Homepage = "https://github.com/joshuasteier/matscope"
Documentation = "https://matscope.readthedocs.io"
Repository = "https://github.com/joshuasteier/matscope"
Issues = "https://github.com/joshuasteier/matscope/issues"

[tool.setuptools.packages.find]
include = ["matscope*"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
